phase_3_llm_integration/requirements.txt
# Phase 3 LLM Integration - Requirements

# Chroma Cloud Client
chromadb>=0.5.0
chromadb[api]==0.5.0

# LLM Integration
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Async support
aiohttp>=3.9.0
asyncio-contextmanager>=1.0.0

# Logging and monitoring
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# API Server (optional)
flask>=3.0.0
flask-cors>=4.0.0

# Database (if using PostgreSQL for metadata)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Performance
orjson>=3.9.0  # Fast JSON serialization

# Development
black>=23.0.0
pylint>=3.0.0
